// src/app/admin/application/types.ts
export interface ApplicationType {
    id: number;
    full_name: string;
    phone: string;
    birth_date: string;
    address: string;
    type: string;
    submitted_at: string;
    application_type: string;
    status: string;
    // Добавьте другие поля, которые вам нужны
} 